refactor(app): inline the disconnect confirm into remote.tsx
The disconnect confirm is a trivial two-button dialog whose deps (Button, Dialog,
useDialog, useLanguage) already ship in the settings page bundle, so giving it its
own lazy module + dynamic import bought nothing. Inline it back into remote.tsx
and drop the separate file and the dynamic import; the heavy connect/pairing flow
stays lazily imported. Disconnect still never loads the pairing state machine.
