+++ packages/app/src/pages/settings/remote.tsx
@@ -1,5 +1,6 @@
 import { Button } from "@opencode-ai/ui/button"
 import { useDialog } from "@opencode-ai/ui/context/dialog"
+import { Dialog } from "@opencode-ai/ui/dialog"
 import { Icon } from "@opencode-ai/ui/icon"
 import { showToast } from "@opencode-ai/ui/toast"
 import { type Component, type ComponentProps, Match, Show, Switch, createSignal, onCleanup, onMount } from "solid-js"
@@ -63,7 +64,10 @@ export const RemotePage: Component = () => {
     )
   }
   const openDisconnect = () => {
-    void import("@/components/dialog-disconnect-remote").then((m) => dialog.show(() => <m.DialogDisconnectRemote />))
+    // Inline, not a lazy module: it is a trivial confirm whose deps are already in
+    // this page's bundle, so a separate dynamic chunk bought nothing. The heavy
+    // connect/pairing flow stays lazily imported.
+    dialog.show(() => <DialogDisconnectRemote />)
   }
 
   const statusLabel = () => {
@@ -172,5 +176,42 @@ function Capability(props: { icon: IconName; text: string }) {
       <Icon name={props.icon} class="size-4 text-fg-weak shrink-0" />
       <span class="text-body text-fg-base">{props.text}</span>
     </div>
+  )
+}
+
+// Irreversible confirm: disconnecting wipes the saved token, so it gets a
+// two-button Dialog with a danger action, per the design system. Inlined here
+// (not a separate lazy module): it is a trivial confirm whose deps already ship
+// with this page, so it never pulls in the connect flow's pairing state machine.
+function DialogDisconnectRemote() {
+  const language = useLanguage()
+  const dialog = useDialog()
+  const [busy, setBusy] = createSignal(false)
+
+  const handleDisconnect = async () => {
+    if (busy()) return
+    setBusy(true)
+    try {
+      await window.api?.remote?.disconnect()
+      dialog.close()
+    } finally {
+      setBusy(false)
+    }
+  }
+
+  return (
+    <Dialog title={language.t("settings.remote.disconnect.title")} fit class="w-full max-w-[420px] mx-auto">
+      <div class="px-6 pt-2 pb-6">
+        <span class="text-body text-fg-strong">{language.t("settings.remote.disconnect.body")}</span>
+      </div>
+      <div class="flex justify-end gap-2 px-6 pb-6">
+        <Button variant="secondary" onClick={() => dialog.close()} disabled={busy()}>
+          {language.t("common.cancel")}
+        </Button>
+        <Button variant="danger" onClick={handleDisconnect} disabled={busy()}>
+          {language.t("settings.remote.action.disconnect")}
+        </Button>
+      </div>
+    </Dialog>
   )
 }
